tweak behavior of pnd override file support
Fix some issues unsetting variables, and .ovr will now be deleted when everything is unset.

That last behavior may cause some issues with .ovr files targetting multiple apps, but libpnd doesn't opperate at a section level nor provide a way to discover which keys are set. We'll see if a revert is coming...